Begin consolidating standard and batch modes of RnaChipIntegrator.
Move the CLI, OutputFiles & AnalysisParams classes and read_feature_file,
read_peak_file & find_nearest_features functions from batch.py to cli.py
and update the 'main' function to use these to implement the standard
RnaChipIntegrator command line program.
